Extract UI logic into DI modules
