+++ src/html/options.html
@@ -24,6 +24,12 @@
       <button id="btn-test-openai" disabled>Test OpenAI</button>
       <p id="status"></p>
     </main>
-    <script type="module" src="../js/options.entry.js"></script>
+    <script type="module">
+      import container from "../js/container.js";
+      document.addEventListener("DOMContentLoaded", async () => {
+        const page = await container.get("GptExt_Ui_Page_Options$");
+        await page.init();
+      });
+    </script>
   </body>
 </html>
